+++ src/main/java/org/broadinstitute/dsde/consent/ontology/Utils.java
@@ -9,8 +9,4 @@ public class Utils {
         return LoggerFactory.getLogger(clazz);
     }
 
-    public static boolean isNullOrEmpty(final String s) {
-        return s == null || s.isEmpty();
-    }
-
 }
