--- src/main/java/org/broadinstitute/dsde/consent/ontology/translate/DTO/RecommendationDTO.java
@@ -1,24 +1,0 @@
-package org.broadinstitute.dsde.consent.ontology.translate.DTO;
-
-public class RecommendationDTO {
-    private String title;
-    private String category;
-
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getCategory() {
-        return category;
-    }
-
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public void setCategory(String category) {
-        this.category = category;
-    }
-}
