+++ src/main/java/org/broadinstitute/dsde/consent/ontology/translate/service/Translate.java
@@ -1,30 +1,22 @@
 package org.broadinstitute.dsde.consent.ontology.translate.service;
 
-import org.broadinstitute.dsde.consent.ontology.Utils;
-import org.broadinstitute.dsde.consent.ontology.translate.DTO.RecommendationDTO;
+import com.github.jsonldjava.shaded.com.google.common.reflect.TypeToken;
+import com.google.gson.Gson;
+import org.broadinstitute.dsde.consent.ontology.translate.DTO.RecommendationDto;
 import org.broadinstitute.dsde.consent.ontology.translate.model.TermItem;
-import org.broadinstitute.dsde.consent.ontology.translate.utils.LoadUtils;
-import org.broadinstitute.dsde.consent.ontology.translate.utils.SearchUtils;
+import org.parboiled.common.FileUtils;
 
 import java.util.HashMap;
 import java.util.List;
 
+import org.apache.commons.lang3.StringUtils;
+
 public class Translate {
 
-    String jsonFilePath;
+    public HashMap<String, RecommendationDto> paragraph(final String paragraph) throws Exception {
+        HashMap<String, RecommendationDto> recommendations = new HashMap<>();
 
-    public Translate(String jsonFilePath) {
-        this.jsonFilePath = jsonFilePath;
-    }
-
-    public HashMap<String, RecommendationDTO> paragraph(final String paragraph) throws Exception {
-        HashMap<String, RecommendationDTO> recommendations = new HashMap<>();
-
-        if (Utils.isNullOrEmpty(paragraph) || Utils.isNullOrEmpty(jsonFilePath)) {
-            return recommendations;
-        }
-
-        List<TermItem> terms = LoadUtils.termsFromLocalResourcesJSON(jsonFilePath);
+        List<TermItem> terms = loadJSONFromResources("search-terms.json");
 
         for (TermItem term : terms) {
             final String title = term.getTitle();
@@ -33,12 +25,13 @@ public class Translate {
             final String[] keywords = term.getKeywords();
 
             for (String keyword : keywords) {
-                final boolean foundMatch = SearchUtils.searchForKeyword(keyword, paragraph);
+                final boolean foundMatch = searchForKeyword(keyword, paragraph);
 
                 if (foundMatch) {
-                    RecommendationDTO recommendation = new RecommendationDTO();
-                    recommendation.setTitle(title);
-                    recommendation.setCategory(category);
+                    RecommendationDto recommendation = new RecommendationDto(
+                            title,
+                            category
+                    );
                     if (!recommendations.containsKey(url)) {
                         recommendations.put(url, recommendation);
                     }
@@ -47,6 +40,23 @@ public class Translate {
         }
 
         return recommendations;
+    }
+
+    public static boolean searchForKeyword(final String keyword, final String targetText) {
+        return StringUtils.containsIgnoreCase(targetText, keyword);
+    }
+
+    public static List<TermItem> loadJSONFromResources(final String jsonFilePath) throws Exception {
+        try {
+            final String searchTerms = FileUtils.readAllTextFromResource(jsonFilePath);
+            return new Gson().fromJson(
+                    searchTerms,
+                    new TypeToken<List<TermItem>>() {
+                    }.getType()
+            );
+        } catch (Exception e) {
+            throw new Exception("Error loading terms from json file: " + jsonFilePath, e);
+        }
     }
 
 }
